Issue四修: #5 -jc检查模式崩溃——arena扩容链缓冲free后页面解约,主函数返空的作用域退出例行release经xt_check_obj读已解约页对象头→0xC0000005;对齐正常路径destroy+终0硬退出(实测s4复现段错误,修复后无错exit0/有错exit1/正常编译三路绿) + #6 运行时平台隔离——xt_self_path补__APPLE__(_NSGetExecutablePath+realpath)/__linux__分支,未知平台#error;通道阻塞三处DWORD/INFINITE→uint32_t+0xFFFFFFFF;POSIX分支XT_CHAN_COND_WAIT返回语义反转修正(与Windows版对齐,原成功返0会被当超时);Windows重建后通道/调度单元(08/09/58/59/71)全绿 + #7 GSC外函数小数返回ABI——C桥一律返回装箱uintptr_t(RAX),旧按注解生declare double读XMM0必错;改为declare i64直取真装箱值 + 渲染链接补-limm32(渲染桥v1.2.x IME的Imm*系API) + #8 GSC zao转译?标识符——新增gi映射(?→_q)覆盖声明预扫描六类/变量/赋值/遍历/参数/类型父子/注入/方法参/同步回写/Identifier引用/模块预扫描与exports与别名/闭包参;模块级外函数预声明补齐(渲染库zao曾卡undefined: XT_*);渲染库zao端到端编译通过

+++ compiler/llvm.go
@@ -3270,7 +3270,12 @@ func (c *LLVMCompiler) compileExternalFunctionStatement(s *ast.ExternalFunctionS
 	case "整", "整数":
 		retType = "i64"
 	case "小数":
-		retType = "double"
+		// C 桥 ABI:浮点外函一律返回装箱 XTValue(uintptr_t,如 渲染桥.c 的
+		// XT_GetFrameTime 返回 xt_make_float(...)),声明必须 i64。旧实现按注解生
+		// declare double,调用结果又被当 XTValue 去 retain → "defined with type
+		// 'double' but expected 'i64'"(Issue#7);且即便过了类型检查,double 走
+		// XMM0、uintptr_t 走 RAX,寄存器读错。与自举版行为对齐。
+		retType = "i64"
 	case "布尔", "逻辑":
 		retType = "i1"
 	case "字", "字符串":
